🎉 Finish project

+++ src/main/java/com/example/Utillity.java
@@ -1,9 +1,9 @@
 package com.example;
 
 public class Utillity {
-    public static double getRunTimeInSeconds(long startTime) {
+    public static double getRunTimeInSeconds(double startTime) {
         long endTime = System.nanoTime();
-        long runTime = endTime - startTime;
+        double runTime = endTime - startTime;
         double runTimeInSeconds = runTime / 1_000_000_000.0;
         return runTimeInSeconds;
     }
